x64

+++ Classes/TL/Variant+NNT.h
@@ -671,7 +671,7 @@ public:
         return *this;
     }
     
-    void set_str(char const* str, core::_bool_copy cpy, uint len = -1)
+    void set_str(char const* str, core::_bool_copy cpy, usize len = -1)
     {
         clear();
